First try at tinder-like swiping
Attempted to create a swipe listener that changes the pages to
instantly go from one empty page, to the other, back to the center
to imitate a looping between two views. Issue: does not loop back to
center. Maybe the animation causes a delay so the switch back to center
doesn't go through.