lock for gps before request, food types, etc
Attempted to create intent to ask for location permission if not allowed
by user, but this is a lot of extra work for now. At least, made network
request wait for GPS data to be received by adding a lock to the request
for network data. Food types retrieved by getting the linear layouts for
left and right columns of categories. Used a stringbuilder to append all
the food categories into one long string.