connectivity receiver, fragment
made a broadcast receiver that is registered with android if the
amazon database task fails to connect. automatically makes a request
to database once network is back. currently works for when we start on
airplane mode and go into app, then turn off airplane mode.

Changed mainpagefragment to be defined in onCreate of tinderactivity, not
in imagePagerAdapter.