no internet for image requests, swiping
made no swiping only occur for when an image is null. no internet message goes away
once internet is re-established, which is a nice way of knowing that loading is
happening. made error function for image async task and set up handleNoInternet to
differentiate between the two tasks and call the appropriate function to continue in
the receiver.