no crash on likes/dislikes, wifi/no wifi check
when user likes/dislikes, those requests will not crash if no internet.
they will not start if no internet at the beginning of the request.
however, if the internet goes down during the request, the dynamoDBMapper
library will constantly repeat its attempt in the background until success.
This should hopefully not affect runtime and be a miniscule load.
in requestfromdatabase, the amount of businesses requested will depend on
if WIFI or some other internet is active.