new gps callback set up
locationListener interface added, code to delegate getting location to
callback set up in handleLocationSettings(), onLocationChanged function
set up. Weird bug where, if the new way is used, the image fragment is
given a drawable but is not visible. the placeholder is still visible, but
other than that, the app acts normally...